user auth

+++ server/exercises-model.js
@@ -1,4 +1,3 @@
-// Import dependencies.
 import mongoose from 'mongoose';
 import 'dotenv/config';
 
@@ -18,35 +17,37 @@ db.once("open", (err) => {
     }
 });
 
-// SCHEMA: Define the collection's schema.
+// Define the collection's schema.
 const exerciseSchema = mongoose.Schema({
     name: { type: String, required: true },
     reps: { type: Number, required: true },
     weight: { type: Number, required: true },
     unit: { type: String, required: true },
-    date: { type: Date, required: true }
+    date: { type: Date, required: true },
+    userId: {type: String, required: true}
 });
 
 // Compile the model from the schema.
 const Exercise = mongoose.model("Exercise", exerciseSchema);
 
 
-// CREATE model
-const createExercise = async (name, reps, weight, unit, date) => {
+// create model
+const createExercise = async (name, reps, weight, unit, date, userId) => {
     const exercise = new Exercise({
         name: name,
         reps: reps,
         weight: weight,
         unit: unit,
-        date: date
+        date: date,
+        userId: userId
     });
     return exercise.save();
 }
 
-// RETRIEVE models
+
 // Retrieve all exercises and return a promise.
-const getExercises = async () => {
-    const query = Exercise.find();
+const getExercises = async (uid) => {
+    const query = Exercise.find({userId : uid });
     return query.exec();
 };
 
@@ -57,26 +58,25 @@ const findExerciseById = async (_id) => {
 }
 
 
-// DELETE model based on ID 
-const deleteById = async (_id) => {
-    const result = await Exercise.deleteOne({ _id: _id });
+// delete model based on ID
+const deleteById = async (_id, uid) => {
+    const result = await Exercise.deleteOne({ _id: _id, userId: uid });
     return result.deletedCount;
 };
 
 
 // REPLACE model
-const replaceExercise = async (_id, name, reps, weight, unit, date) => {
+const replaceExercise = async (_id, name, reps, weight, unit, date, uid) => {
     const result = await Exercise.replaceOne({ _id: _id }, {
         name: name,
         reps: reps,
         weight: weight,
         unit: unit,
-        date: date
+        date: date,
+        userId: uid
     });
     return result.modifiedCount;
 }
-
-
 
 // Export our variables for use in the controller file.
 export { createExercise, getExercises, findExerciseById, replaceExercise, deleteById }
